update import statements after moving handlers to directory
